int to long objectid

+++ src/main/java/com/manyi/mqservice/service/MessageServer.java
@@ -155,7 +155,7 @@ public class MessageServer {
 	try {
 	    String msgTopic = (String) msg.get("msgTopic");
 	    String msgType = (String) msg.get("msgType");
-	    int objectId = Integer.valueOf(msg.get("objectId") + "");
+	    long objectId = Long.valueOf(msg.get("objectId") + "");
 	    int userId = Integer.valueOf(msg.get("userId") + "");
 	    String memo = (String) msg.get("memo");
 	    String ip = msg.get("ip") + "";
